requirements.txt
# ==============================================================================
# requirements.txt — Runtime + Test dependencies
# ==============================================================================

# LangGraph / LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-ollama>=0.2.0

# HTTP / Data
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Config / Utilities
python-dotenv>=1.0.0
typing-extensions>=4.8.0

# HuggingFace (for src/download.py)
huggingface_hub>=0.20.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
requests-mock>=1.11.0
